+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.17.2
+
+### ✨ Features
+
+- fix: download attachment with the filename @RETFU (#957)
+
+### ⛓️ Dependency
+
+- Bump mail from 0.3.1 to 0.4.0 @dependabot (#958)
+- Bump ex\_aws from 2.5.4 to 2.5.5 @dependabot (#952)
+- Bump plug\_cowboy from 2.7.1 to 2.7.2 @dependabot (#953)
+
 ## 1.17.1
 
 ### 🐛 Bug Fixes